requirements.txt
datamaps
-e git+https://github.com/banillie/projectlibrary#egg=projectlibrary
python-docx==0.8.10
openpyxl==3.0.0
matplotlib

